Improve time to show ads (BOTTOM BANNER ONLY)
Changing the request to be sent only the first time the ad is displayed, improved the time to show.
Send request with the ad hidden will discard the received and change update time to 60 sec, its how ad mob works. Keeping the adview hidden for too long also will trigger this condition.